requirements.txt
Flask
textblob
psycopg
nltk
